backend/app/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv
import os
from pymongo import MongoClient

# --- load env vars (.env file in backend root) ---
load_dotenv()

MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://localhost:27017")
client = MongoClient(MONGODB_URI)
db = client.namattek_booking   # we'll use later

app = FastAPI(title="NamatTech Booking API")

# allow the future React front‑end to call this API
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/healthz")
def healthz():
    return {"status": "ok"}

from .routes import services, customers, appointments

app.include_router(services.router)
app.include_router(customers.router)
app.include_router(appointments.router)

