backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
fastapi-scheduler
python-dateutil
